atomic design applied => app structure & components updates
