--- src/components/Bio.js
@@ -1,24 +1,0 @@
-export default function Bio() {
-	return (
-		<div className="col-md-4" style={{ paddingTop: "70px" }}>
-			<div className="card">
-				<img
-					className="card-img-top"
-					src="img/vadim-gerko-zdjecie-cv.jpg"
-				/>
-				<div className="card-body">
-					<p>
-						<strong>Hi there!</strong> I'm a self-taught React /
-						Firebase Developer building full-stack web apps that
-						help people to push their (and mine) creativity &
-						productivity to its limits! Check out my portfolio for
-						more info & projects!
-					</p>
-				</div>
-				<div className="card-footer">
-					<p className="text-muted">Lublin, Poland</p>
-				</div>
-			</div>
-		</div>
-	);
-}
