+++ src/components/organisms/Footer.js
@@ -1,0 +1,18 @@
+export default function Footer() {
+	return (
+		<footer>
+			<hr />
+			<p className="text-center">
+				designed by{" "}
+				<a
+					href="https://github.com/vadimgierko"
+					target="_blank"
+					rel="noreferrer"
+				>
+					Vadim Gierko
+				</a>{" "}
+				&copy; 2021-2022
+			</p>
+		</footer>
+	);
+}
